pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64"]

[project]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering"
]
dependencies = [
  "numba>=0.64.0",
  "numpy>=2.4.2",
  "optv>=0.3.2",
  "pyyaml>=6.0.3",
  "scipy>=1.17.1"
]
description = "Python version of the OpenPTV library"
license = {file = "LICENSE"}
name = "openptv-python"
version = "0.5.0"
readme = "README.md"
requires-python = ">=3.11,<3.14"

[project.scripts]
pyptv = "pyptv.pyptv_gui:main"

[project.optional-dependencies]
docs = [
  "myst-parser>=5.0.0",
  "pydata-sphinx-theme>=0.16.1",
  "sphinx>=8.0.0,<9.1.0",
  "sphinx-autoapi>=3.7.0"
]
gui = [
  "traits>=6.4.0",
  "traitsui>=7.4.0",
  "enable>=6.1.0",
  "chaco>=6.1.0",
  "PySide6>=6.0.0",
  "scikit-image>=0.20.0",
  "pandas>=2.0.0",
  "matplotlib>=3.7.0",
  "tables>=3.8.0",
  "tqdm>=4.65.0",
  "imagecodecs>=2023.1.23",
  "flowtracks>=0.3.0",
  "Pygments>=2.15.0",
  "pyparsing>=3.0.0",
  "marimo>=0.19.10"
]
test = [
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0"
]
dev = [
  "mypy>=1.19.1",
  "myst-parser>=5.0.0",
  "pre-commit>=4.5.1",
  "pydata-sphinx-theme>=0.16.1",
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0",
  "sphinx>=8.0.0,<9.1.0",
  "sphinx-autoapi>=3.7.0",
  "types-pyyaml>=6.0.12.20250915",
  "traits>=6.4.0",
  "traitsui>=7.4.0",
  "enable>=6.1.0",
  "chaco>=6.1.0",
  "PySide6>=6.0.0",
  "scikit-image>=0.20.0",
  "pandas>=2.0.0",
  "matplotlib>=3.7.0",
  "tables>=3.8.0",
  "tqdm>=4.65.0",
  "imagecodecs>=2023.1.23",
  "flowtracks>=0.3.0",
  "Pygments>=2.15.0",
  "pyparsing>=3.0.0",
  "marimo>=0.19.10"
]

[tool.coverage.run]
branch = true

[tool.mypy]
strict = false

[tool.ruff]
# Same as Black.
indent-width = 4
line-length = 88

[tool.ruff.lint]
ignore = [
  # pydocstyle: Missing Docstrings
  "D1"
]
select = [
  # pyflakes
  "F",
  # pycodestyle
  "E",
  "W",
  # isort
  "I",
  # pydocstyle
  "D"
]

[tool.ruff.lint.pycodestyle]
max-line-length = 110

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.setuptools]
packages = ["openptv_python", "pyptv"]

[tool.setuptools.package-dir]
openptv_python = "src/openptv_python"
pyptv = "src/pyptv"

